Refactor transaction handling to support pagination in ListTransaction use case. Update transaction amount type from int64 to float64 for better precision. Enhance transaction DTO validation by removing redundant checks. Implement fetching transactions by account number in the repository.

+++ payment/infrastructure/projection/persistent_object/transaction.go
@@ -8,6 +8,6 @@ type Transaction struct {
 	TransactionCode string `gorm:"unique"`
 	AccountNo       string
 	Type            constant.TransactionType // Deposit, Withdraw, Transfer
-	Amount          int64
+	Amount          float64
 	Reference       string // optional
 }
